added directionsCardFloor created
